flora_synthesis/requirements.txt
django>=4.2
gunicorn
whitenoise
django-cors-headers
psycopg2-binary
requests
beautifulsoup4
djangorestframework
django-cors-headers
factory_boy
django-debug-toolbar
requests-ratelimiter
django-q2
flake8
flake8-docstrings
